third_party/WebKit/LayoutTests/fast/dom/Element/scrollTop-scrollLeft-frameset-expected.txt

Test for scrollTop/Left values of document.body and document.documentElement in documents where the body is a frameset element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.scrollingElement is document.documentElement
PASS iframeQuirks.contentDocument.body is iframeQuirks.contentDocument.querySelector('frameset')
PASS iframeQuirks.contentDocument.scrollingElement is null
PASS iframeQuirks.contentDocument.documentElement.scrollTop is 0
PASS iframeQuirks.contentDocument.documentElement.scrollLeft is 0
PASS iframeQuirks.contentDocument.body.scrollTop is 0
PASS iframeQuirks.contentDocument.body.scrollLeft is 0
PASS iframeQuirks.contentWindow.pageYOffset is 500
PASS iframeQuirks.contentWindow.pageXOffset is 200


Add extra body element
PASS iframeQuirks.contentDocument.body is iframeQuirks.contentDocument.querySelector('frameset')
PASS iframeQuirks.contentDocument.scrollingElement is bodyElement
PASS iframeQuirks.contentDocument.documentElement.scrollTop is 0
PASS iframeQuirks.contentDocument.documentElement.scrollLeft is 0
PASS iframeQuirks.contentDocument.body.scrollTop is 0
PASS iframeQuirks.contentDocument.body.scrollLeft is 0
PASS iframeQuirks.contentWindow.pageYOffset is 500
PASS iframeQuirks.contentWindow.pageXOffset is 200
PASS iframeQuirks.contentDocument.scrollingElement.scrollTop is 500
PASS iframeQuirks.contentDocument.scrollingElement.scrollLeft is 200
PASS successfullyParsed is true

TEST COMPLETE
